add resume and history related functions; reorganize workflows code
